* bug
git-svn-id: https://svn.shadanakar.org/onPHP/trunk@926 db88c99c-74ec-0310-8720-8416fdbde762

+++ main/DAOs/SmartDAO.class.php
@@ -50,11 +50,13 @@
 			$cache = Cache::me();
 			
 			$indexList = $cache->mark($className)->get($indexKey);
-			
-			$cache->mark($className)->delete($indexKey);
-			
-			foreach ($indexList as $key => &$true)
-				$cache->mark($className)->delete($key);
+
+			if ($indexList) {
+				$cache->mark($className)->delete($indexKey);
+				
+				foreach ($indexList as $key => &$true)
+					$cache->mark($className)->delete($key);
+			}
 			
 			return true;
 		}
